test: introduce obsidian test with wdio
