+++ package.json
@@ -14,7 +14,8 @@
     "postbump": "dotenv -- if-env SKIP_POST_BUMP=true && echo skip bump || git push --follow-tags origin master",
     "predeploy": "bun run bump",
     "deploy": "bun run export",
-    "lint": "bun biome format --write src/"
+    "lint": "bun biome format --write src/",
+    "e2e": "bun wdio run ./wdio.conf.mts"
   },
   "engines": {
     "node": "^22.2.0"
@@ -26,6 +27,10 @@
     "@biomejs/biome": "^2.0.0-beta.2",
     "@types/bun": "^1.2.12",
     "@types/node": "^22.15.3",
+    "@wdio/cli": "^9.12.7",
+    "@wdio/local-runner": "^9.12.7",
+    "@wdio/mocha-framework": "^9.12.6",
+    "@wdio/spec-reporter": "^9.12.6",
     "builtin-modules": "5.0.0",
     "esbuild": "0.25.3",
     "i18next": "^25.0.2",
@@ -36,14 +41,18 @@
   "dependencies": {
     "@delucis/if-env": "^1.1.2",
     "@enveloppe/obsidian-dataview": "^0.5.70-2",
+    "@types/mocha": "^10.0.10",
     "ansi-colors": "^4.1.3",
     "commander": "^13.1.0",
     "commit-and-tag-version": "^12.5.1",
     "dedent": "^1.6.0",
     "dotenv": "^16.5.0",
     "dotenv-cli": "^8.0.0",
+    "mocha": "^11.2.2",
     "obsidian-typings": "^3.5.1",
-    "uniformize": "^2.4.0"
+    "uniformize": "^2.4.0",
+    "wdio-obsidian-reporter": "^1.2.0",
+    "wdio-obsidian-service": "^1.2.0"
   },
   "trustedDependencies": [
     "@biomejs/biome",
