+++ tests/fixtures/simple_add.md
@@ -1,0 +1,4 @@
+---
+---
+
+foo:: bar
